refactor: remove unused packages and simplify architecture
- Remove empty directories: api/, blockchain/, consensus/, domain/, network/, persistence/
- Remove unused packages: crypto/, service/, config/, node/
- Remove CommonUtil.kt (unused utility functions)
- Move Policy.kt to core package and simplify
- Update imports in Router.kt and API.kt
- Remove unused encodeToString import from lightnode Application.kt
- Delete obsolete REFACTORING_GUIDE.md
- Update README.md to reflect simplified 3-package structure (core/, state/, util/)
